bugfix: startup of disable-short-notation (Luigi)

+++ dgt/display.py
@@ -134,7 +134,7 @@ class DgtDisplay(DisplayMsg, threading.Thread):
                 text = Dgt.DISPLAY_MOVE(move=self.last_move, fen=self.last_fen, side=side, wait=False, maxtime=1,
                                         beep=beep, devs={'ser', 'i2c', 'web'}, uci960=self.uci960,
                                         lang=self.dgttranslate.language, capital=self.dgttranslate.capital,
-                                        short=self.dgttranslate.notation)
+                                        long=self.dgttranslate.notation)
             else:
                 text = self.dgttranslate.text('B10_nomove')
             DispatchDgt.fire(text)
@@ -186,7 +186,7 @@ class DgtDisplay(DisplayMsg, threading.Thread):
                 text = Dgt.DISPLAY_MOVE(move=self.hint_move, fen=self.hint_fen, side=side, wait=False, maxtime=1,
                                         beep=beep, devs={'ser', 'i2c', 'web'}, uci960=self.uci960,
                                         lang=self.dgttranslate.language, capital=self.dgttranslate.capital,
-                                        short=self.dgttranslate.notation)
+                                        long=self.dgttranslate.notation)
             else:
                 text = self.dgttranslate.text('B10_nomove')
             DispatchDgt.fire(text)
@@ -493,7 +493,7 @@ class DgtDisplay(DisplayMsg, threading.Thread):
         disp = Dgt.DISPLAY_MOVE(move=move, fen=message.game.fen(), side=side, wait=message.wait, maxtime=0,
                                 beep=beep, devs={'ser', 'i2c', 'web'}, uci960=self.uci960,
                                 lang=self.dgttranslate.language, capital=self.dgttranslate.capital,
-                                short=self.dgttranslate.notation)
+                                long=self.dgttranslate.notation)
         DispatchDgt.fire(disp)
         DispatchDgt.fire(Dgt.LIGHT_SQUARES(uci_move=move.uci(), devs={'ser', 'web'}))
         self.leds_are_on = True
@@ -571,7 +571,7 @@ class DgtDisplay(DisplayMsg, threading.Thread):
             disp = Dgt.DISPLAY_MOVE(move=self.hint_move, fen=self.hint_fen, side=side, wait=True, maxtime=0,
                                     beep=beep, devs={'ser', 'i2c', 'web'}, uci960=self.uci960,
                                     lang=self.dgttranslate.language, capital=self.dgttranslate.capital,
-                                    short=self.dgttranslate.notation)
+                                    long=self.dgttranslate.notation)
             DispatchDgt.fire(disp)
 
     def _process_startup_info(self, message):
@@ -635,7 +635,7 @@ class DgtDisplay(DisplayMsg, threading.Thread):
                     text = Dgt.DISPLAY_MOVE(move=self.hint_move, fen=self.hint_fen, side=side, wait=True, maxtime=1,
                                             beep=beep, devs={'ser', 'i2c', 'web'}, uci960=self.uci960,
                                             lang=self.dgttranslate.language, capital=self.dgttranslate.capital,
-                                            short=self.dgttranslate.notation)
+                                            long=self.dgttranslate.notation)
                 else:
                     text = self.dgttranslate.text('N10_nomove')
             else:
@@ -656,7 +656,7 @@ class DgtDisplay(DisplayMsg, threading.Thread):
             beep = self.dgttranslate.bl(BeepLevel.BUTTON)
             text = Dgt.DISPLAY_MOVE(move=self.play_move, fen=self.play_fen, side=side, wait=True, maxtime=1,
                                     beep=beep, devs=devs, uci960=self.uci960, lang=self.dgttranslate.language,
-                                    capital=self.dgttranslate.capital, short=self.dgttranslate.notation)
+                                    capital=self.dgttranslate.capital, long=self.dgttranslate.notation)
         else:
             text = None
             if self._inside_main_menu():
